[coverage] Disable hwasan globals on clang_coverage_test_bin
This particular target fails to link because a tagged version of
__llvm_profile_counter_bias provided in clang_coverage_test.cc causes
an overflow check to fail when resolving a relocation in an
unistrumented profile runtime. For now, we can just disable
hwasan-globals for this specific target. This is done through a config
so we guarantee the flag comes at the very end of the compile
invocation.

Bug: 107896
Change-Id: I8f1ce07d6b945ffa410a6e3862ac16e263261938
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716632
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>