[element_manager]: don't drop element til closed
ElementManager::run_element_until_closed() blocks until the element is
closed. This function is called as an async task, and the function
(and therefore the task) takes ownership of the `element` instance.
This way, the `element` instance is not dropped until it is considered
closed.

For elements based on CFv2 components, dropping the element does not
appear to cause an issue. But elements for CFv1 components hold onto an
`App`, and the `App` holds a `controller`, and the `controller` "Keeps
the component alive until `App` is dropped."

A recent change removed the `element` from the call to
`run_element_until_closed()` because the function no longer needed the
`element`. Tests against components migrated to CFv2 still worked. But
there is an end-to-end test that has not yet been migrated, and the test
now fails because the `element` is no longer passed into the task.

This CL reverts the part of the change that removed the element, so the
element will no longer be dropped until the function completes.

Fixed: 108032
Change-Id: Iec0e5fa38c93ec1bd7ea5b87d1c920d4df779c4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717362
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Rich Kadel <richkadel@google.com>
Reviewed-by: Shai Barack <shayba@google.com>