[fxfs] Fix volume termination race
It was possible to race closing of a file against locking a store.  If
the file was unlinked, this could trigger a tombstone which would trip
assertions if the store was locked.

This CL fixes the issue by ensuring the pager is properly terminated
before attempting to lock the store.

This CL also:

  + adds new lock states so that we can more easily diagnose these
    issues in future.

  + ensures that we do the initial reap of encrypted object stores.
    This was missing.

  + allows unlocking volumes when the filesystem is read-only.

Fixed: 107470
Fixed: 107695
Change-Id: I2a92a00a6f91f994b871d67cfa106beaab8119af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715789
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Chris Suter <csuter@google.com>