[f2fs] Prepare for componentizing f2fs
This CL introduces Runner class for non-component operation and
breaks the vfs object out into the Runner object which creates
and owns the f2fs object as blobfs and minfs do. This way allows
f2fs to be able to handle a startup service before starting the
filesystem service when it runs as a component.

Bug: 107157

Change-Id: Ifc0ac1d07a17c360944c0d37c8a6b5d3501045f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716963
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Seokhwan Kim <sukka.kim@samsung.com>