[driver_runtime] Protocol discovery - call service_connect op.
This is a temporary implementation using v1 of the protocol discovery
API. The child driver tries to connect to a driver runtime protocol
using DdkServiceConnect, however in the compat-shim this tries to
make a FIDL RuntimeConnector::Connect call to the parent. We need
to implement this protocol if the parent driver is running the compat
shim.

This is so the wlan driver can work in compat mode in DFv2. We should
replace the API with the newer protocol discovery when
available.

Bug: 107866
Change-Id: I9512c7759ae85e2520b3cbc55ae42e2dd36e8889
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716819
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>