[factoryfs] Update TODOs and unsupported responses.
Replacing TODOs to point to the bug expected to handle them. Leaving old
TODOs for now that we don't expect to actually handle for now.

Replacing NOT_SUPPORTED responses with NOT_FILE as appropriate in
directory class. Not bothing with file class as it should shift to
inheriting from vmo_file in order to support GetBackingMemory().

Bug: 105904
Change-Id: I9d11dada41e0f763b814818f8b826f5cec73586c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717773
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>