[cts] Impl more policies for JSON schema changes.
Changes to the value of an "enum" key are now classified as breaking
changes. Changing the value of "additionalProperties" to True is
non-breaking, but to False is breaking. Any breaking change that doesn't
come with a change to the "id" field is another error.

Change-Id: Ieafcab27b7ea10144df2df77265a26677c0251e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715689
Reviewed-by: Victor Chang <vfcc@google.com>
Commit-Queue: Isabelle Goralsky <igoralsky@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>