[netstack3] Move create listener into datagram.rs
Move the functionality of creating listener sockets into the datagram
module and delegate to that when allocating a UDP listener.

This also has the advantage of removing the linear scan that was being
performed when allocating a socket for a UDP listener since the same
information is available in constant time from the socketmap.

Bug: 104462
Change-Id: I61af1973dbc5db8d069cb676179d130fb921b7f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714620
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>