[intel-i915-tgl] Remove WAIT_ON*() macros.
The macros are replaced with a templated function, and the wait
condition is expressed as a lambda argument. This aligns with the C++
style guide, which states that preprocessor macros should only be used
where there is no alternative.

Tested: fx test intel-i915-tgl-test.cm
Change-Id: I9356607fd6be88eeea8c815579090075b49f866f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717563
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>