[virtualization][vsock] Re-poll on spurious wakeup
It's possible for a single-reader to get one spurious wakeup (between
times when there's actually data on the socket). If there's a spurious
wakeup, this device should just re-poll.

Tested:
Ran the perf test and confirmed that A) there is at most one spurious
wakeup and B) we no longer waste a chain on spurious wakeups

Bug: 97355
Change-Id: If62b9b744e8f24e3c0c54e85f2ffa98a0854a331
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717583
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>