[nxpfmac] Create an event handler
In order to handle events from mlan we need a centralized event handler
that allows for registering of event callbacks. This creates a very
basic event handler that allows for registering event callbacks per
interface or for all interfaces or events that are not associated with
an interface.

Test: Included unit test
Change-Id: Ia03be61a2bf8c88eefb8d04eaf90c37f648b267e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708983
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>