[virtualization][vsock] Run the util in its own thread
At the moment we're running the util in the main thread and the test in
a new thread which is explicitly disallowed (can't pass FIDL pointers
between threads).

This swaps to running the util in a new thread. This is a bandaid to the
larger issue of this test harness not being thread safe to unblock
some vsock work.

We'll need to re-visit allowing non-blocking commands in the future.

Change-Id: Ia7bcb342b163737d5e8afbdb1ab9cc8ee1aa0abe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717591
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>