[settings] Add inspect structure for external apis
To log external service events to inspect, the service context's
ExternalServiceProxy will be used to handle ExternalServiceEvents from
a common point. This introduces the event and structure for doing so,
and converts direct calls on the ExternalServiceProxy to use the call
and call_async macros, in order to log the external request.

Bug: 94361
Test: fx test setui_service_tests

Change-Id: I28c1a7a64e41cac433ecc262fdbadbe26c3c6016
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705626
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>