[virtualization] Merge linux_runner with termina_guest_manager
This updates the code in //src/virtualization/bin/linux_runner to link
with the //src/virtualization/bin/guest_manager code. We then use a
pointer to the GuestManager from the linux_runner instead of connecting
over FIDL.

This is just a mechanical change and no improvements are added yet. For
example, starting the vm using the TerminaGuestManager protocol does not
yet wire up the needed vsock services, but we will add these
improvements in a following CL. We also don't yet move or rename the
linux_runner code to minimize the noise in the CL.

The effect of this CL is we produce the component
'termina_guest_manager.cm' out of //src/virtualization/bin/linux_runner
and delete the former, standalone termina_guest_manager package.

Bug: 107399
Change-Id: Iaeebeeaf5960df0d06bb374f3045311bc8dab3db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710344
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>