[appmgr] Extract loopback capabilities from appmgr core shard
Some capabilities are defined as routes through `/core` in the
appmgr core shard.

To move towards an appmgr-less world, we need to extract the
capabilities that have nothing to do with appmgr anymore and
put them in a separate shard, so that they can be handled
separately.

Bug: 108213
Change-Id: Idb3913b447fffc1a4d427ddf327445ab8abec30e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718343
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>