[lowpan-ot-driver] Reset OpenThread stack if busy for longer than 30 sec
Scans are somewhat blocking operations---the device cannot actively
participate on the network while one is in progress. Occasionally we can
run into bugs like <fxbug.dev/106509>, where the scan never finishes.
Because there is no way to cancel an ongoing scan in OpenThread, the
only way to get ourselves out of this state is to reset OpenThread.

This CL introduces a `scan_watchdog` which monitors for when the scan
starts and then makes sure that it lasts no longer than
`SCAN_WATCHDOG_TIMEOUT`. If it does, it will terminate the loop, which
will cause OpenThread to be reset.

Bug: 108271
Bug: 106509
Test: Manual verification.
Change-Id: Ie25671770b29bc83ddc9a87ebe7e85a745420ecd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718442
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>