[ffx][config] Add isolation of env vars.
This adds environment variables to the set of things that are (or at
least can be) isolated. There are some areas this doesn't touch yet,
like the emulator library (which has its own env isolation library),
preflight, and self-test. Those will be adapted to this in future
changes.

Fixed: 107979
Change-Id: Iddb4b5495d14c87cb4f8c6f5551a73eb55e7f54e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707808
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>