[cts] Update SDK directory CT error message.
SDK directory layout compatibility test is currently not in the local
build graph. When CQ fails, the developer needs args to rebuild locally
and reproduce the error in order to fix it. Also, combined the golden
generation and CT target lists since the CT is dependent on the golden
files being generated.

Change-Id: Id179757f8613acf36da3649f5cb6fdda82a20505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717951
Commit-Queue: Isabelle Goralsky <igoralsky@google.com>
Reviewed-by: Victor Chang <vfcc@google.com>