[fuchsia-tpm] Implements fuchsia.tpm namespace.
This library intends to implement the fuchsia.tpm namespace
so that it can be shared between the cr50_agent and the tpm_agent.
This allows variation by board type where the cr50_agent is a superset
of the functionality in the tpm_agent. By placing this code in a shared
location it prevents code duplication / drift between the two components.
This initial implementation implements just transmit. Both deprovision
and provision first require our TSS implementation to be ported to Rust.

Bug: 10528

Change-Id: Ife950f70793d4717f2dff89dcf22802d40e68627
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712387
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Fuchsia-Auto-Submit: Ben Wright <benwright@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>