[testrunner] Properly exit package resolving loop if task finished.
Before, we'd wait for the function to finish but it actually goes
through all the package urls and does a best effort to resolve, meaning
it doesn't exit if one fails. In the case that it fails due to a
connection issue, we try to reconnect for 2 minutes and we repeat that
for each url. While the loop does break if the context gets canceled,
the context doesn't get canceled until the function returns or we hit
the botanist timeout. So we should have a separate context for the
package resolving function and cancel it when execute() is finished.

Change-Id: Ibad8891f996fc7258e1877bb630300b14abccf74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718007
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>