[virtualization] Use structured config in termina_guest_manager
There are a number of configuration options to the termina_guest_manager
that are augmented by setting preprocessor symbols and recompiling the
target sources. This is difficult to test since code may be #ifdef'd out
and won't always compile.

This CL introduces some structured configuration[1] to
termina_guest_manager so that we can still control behavior using GN
args, but we also remove the need to #ifdef code. Additionally
structured configurations are much easier to change within test
contexts.

This CL also removes the vestegial support for using a prebuilt stateful
partition. This was useful when network support was not reliable enough
to bootstrap the container but hasn't been used in recent months.

Bug: 107399
Change-Id: Ib18c3925c1fc02ae8c4839635e3165f92c672181
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714871
Reviewed-by: Rupert Shuttleworth <ruperts@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>