[virtualization] Stop providing default values in the CLI.
Instead we want to let the target guest package determine the defaults
for these. Specifically consider an option sucha as virtio-gpu, which was
previously a 'default on' option. If we launch termina using the CLI, we
do not want there to be a GPU device added automatically by the CLI.

In other words, the CLI would not mutate any option of a guest package
unless it is specifically requested as an override.

Bug: 107399
Change-Id: Ic46747c60d653a9051a9a4b1ed93fe7fa75c1c34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711426
Reviewed-by: Alexander Radchenko <alerad@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>