[installer] Speed up installer
This speeds up the installer from ~30m to ~4m when installing
workstation onto an Atlas.

We use RemoteBlockClient to read the FVM image from the installer drive.
Reading larger chunks at a time from the USB drive is responsible for
the majority of the speed up.

Bug: 107813
Change-Id: I13c278b547e60a343279ba85b37edb6bd2a75df2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718776
Commit-Queue: Adam Chyb <achyb@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>