[starnix] Always handle special errnos in dequeue_signal
Syscalls can return special errno value to let the kernel restart them
in case of interruption.

Starnix handled these in dequeue_signal where is replaces the rax with
either the EINTR errno value, or with the right value to ensure a
restart. But it only does this if a signal is queued. Otherwise, it let
the errno value as it is, which is incorrect, as user space should never
see these values.

This CL updates dequeue_signal to always handle the special values,
whether there is a signal queued or not.

Fixed: 107926

Change-Id: I25c2dd10c641782f6f8a069a989d694bce7eaa9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718843
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>