[bt][avrcp][a2dp] Subscribe to player changed notification events.
When an a2dp sink, we should detect when the available players have
changed, query to find the new players to see if there are any that are
active. If none of the players are active, we change the player state to
idle.

We also should detect when the addressed player has changed. This event
will be treated similar to track change event and should trigger
attribute update.

Bug: 103847
Test: fx test -v -o bt-avrcp-tests && fx test -v -o bt-a2dp-tests
Change-Id: I163b5d2205519ca9c23f56fc1984e68b820cc0c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703168
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
API-Review: Marie Janssen <jamuraa@google.com>