[minfs] Fix shutdown race with periodic sync.
If the periodic sync is in the middle of running during destruction,
there can be a lot of havoc. This will now try to dequeue any waiting
sync calls, and if it can't find one to dequeue, it ensures that one
isn't running by pushing a task into the queue and blocking on it.

Removes vestigial RefCounted interface and marks minfs final so that
we have full control over destructor ordering.

Bug: 107070
Change-Id: I2ea5e94a83b32fa7f027264dcea4bf87b1344e8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714069
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>