[mock-ddk] Use syslog and add API for configuring global log severity
Previously, mock-ddk logs did not emit the message severity, which meant
that tests with ERROR logs would pass because the test harness couldn't
detect them. Now that the logs include the severity, we have to opt-in a
bunch of passing tests with expected ERROR logs.

Bug: 85835
Change-Id: If626ea47fa3f7a4cf47966576b6e198891c51793
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718222
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Ian Fisher <iafisher@google.com>