[fuchsia-tcti] Add a Fuchsia TCTI Interface.
The TCG TSS 2.0 TPM Command Transmission Interface (TCTI) is the
lowest layer in the TPM Software Stack (TSS). It is a trivial
interface that defines how bytes are sent to and recieved from
the TPM. This API must be C compatible as its lifetime and
calling structure is managed by the TSS 2.0 library which we
imported in an external repository. This CL simply creates
a small internal context which opens a FIDL connection
to the tpm_agent or cr50_agent depending on board type
and uses the fuchsia.tpm.transmit.TransmitCommand interface
to send  commands to the TPM. Responses are recorded on
the heap in a recv_buffer and data is removed from that
buffer as calls occur to fuchsia_tpm_recv.

Bug: 105028
Change-Id: Ie24cd157c6b01362746260d0995b7a858a4b7933
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714016
Commit-Queue: Ben Wright <benwright@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>