[fidl][cpp] Make fidl::WireRequest just an alias
Now that we have predictable, somewhat sane ways to name the types that
represent method arguments we can stop using fidl::WireRequest<M> to be
a special type for requests - we can just use the actual underlying
type.

It can still be useful to have a way to get the request type for a given
method so this leaves fidl::WireRequest as that, but it's just an alias
based on some special generated traits stuff.

Bug: 101525
Change-Id: I59e0ccf0c70392efed739fab8a0bc25cd40d50ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718644
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>