[emu] Move port assignment to later in startup
Previously, port mapping was handled as part of the "make_configs" step
in the start-up process. Any ports that are not explicitly mapped by the
user are assigned available ports by the system.

Because this took place before staging, there is a race condition
possible where an available port is selected, then while staging takes
place the same port is grabbed by another process. When the emulator
launches, it fails to bind to the selected port and start-up is
terminated. See the attached bug.

This CL shifts port assignment from make_configs to the very end of the
staging process. At this point, all filesystem work before the child is
launched have taken place (unless the user chooses to manually edit the
config).

Note that this doesn't completely remove the race condition described
above; it only reduces the window where another process could subvert
the assignment.

This is the latest point in the start-up sequence I can find to assign
ports with our current capabilities. It would be possible to create a
template helper to handle port assignment after manual editing, but
doing so would leave "None" values in the engine.json file, so the user
would have no way to determine which port values were mapped post-facto.
The QMP offers no way to view existing port assignments in the QEMU
versions we include with the SDK, so that option is unavailable.

Bug: 103540

Change-Id: I93bc1173913eb6f699a20a19d0428388854afc9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718045
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Chong Gu <chonggu@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>