[ffx] Improve clarity of the plugins file template
Rewrites the logic blocks in the plugins.md file to be easier to
comprehend, with fewer layers of indirect logic and clear exit points to
the generated functions.

This also renames the template files to not imply that they're
documentation, but rust jinja templates. All references I could find to
them were also changed. I'm not sure why they were given .md extensions
to begin with, but now editors will highlight them properly and cq
(hopefully) won't treat them as no-ops for CI purposes.

Note that this reveals a potential flaw in the `ffx schema` output that
might need to be resolved, as replicating the prior logic results in
some output being masked (a ProfileMemoryOutput on `ffx profile memory`
is not included with this logic, but would be with logic that didn't
only print out the plugin's output if there are no sub-plugins.

However, it's not clear how that should be output, the format of this
does not lend itself to adding leaf items under branch nodes.

Change-Id: Ic459e1551ba80e59c304508ca62e8fa3dc28aec1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718424
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>