[virtualization][vsock] Fix client initiated shutdown race
The current behaviour of the device is to send a SHUTDOWN_BOTH packet
from the read write state, but the guest can reply to it (with a RESET)
before the state transition finishes which sends the connection to a
forced shutdown state.

The read write state should only send half shutdown packets (since it
can still read or write on the other half), and let the client initiated
shutdown state send the full shutdown.

Tested:
1) Added a 1s delay between sending shutdown and transitioning between
   states
2) Ran the Termina vsock integration test
3) Confirmed that the device can mistakenly enter forced shutdown (which
   sends reset packets back to the guest when it thinks the shutdown was
   clean)

Bug: 97355,107486
Change-Id: If8cacb305adb6a550196dc21050bd8394be9dd14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718084
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>