[debugger] Redirect debug_agent's log to zxdb
Provides a general mechanism to allow the debug_agent to display logs in
zxdb. This is required by the TestLauncher because the error is only
available asynchronously. It's also useful to display info and debug
logs in one place for debugger development in the future.

Change-Id: I47ce8874a88defeb5571403534cc9ce081ec334f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716803
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>