[kernel][time] Save timer offset on suspend.
This ensures we do not get time goes backwards errors on resume. It only
affects the get_ticks syscall. This is a short term measure so that we
can test suspend and resume. In the longer term a revamp of the vDSO is
planned so that ticks can be offset in a more holistic manner.

Bug: 89735
Bug: 91701

Change-Id: I095163cc612eeb14250c7e109227d730216efaed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698004
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>