[microbenchmarks] Add same-CPU and different-CPU variants of a round-trip test
Measure the IPC round trip time when the threads involved are pinned
to the same CPU or different CPUs.

The motivation is to investigate how factors like lock contention and
IPI latency affect the different-CPU case compared with the same-CPU
case.

Add the plumbing for making this work for both the single-process and
multi-process cases.  The plumbing for the thread affinities is a bit
fiddly.  The parent thread affinity needs to be set via a wrapper for
the test, while the child thread affinity needs to be passed through
the test into the new process or thread.

I checked manually, by looking at traces, that the threads do actually
get scheduled on the same or different CPUs.

Bug: 45802
Change-Id: I91adf7fabde69345b275a1b78e9a91cd381e117b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718443
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>