[input][shortcut] Implements fuchsia.ui.shortcut2
Shortcut2 is a simplified rework of fuchsia.ui.shortcut.

However, it offers more flexibility for defining shortcuts
that depend on keymaps instead of hardware keys, allowing
all keyboards, including virtual ones, to support
shortcuts.

This is not a full implementation - a11y provisions are
deferred for the time being, and will be backfilled in
due time. Also, no special provisions for non-Latin
shortcuts, but the functionality can be added in a
straightforward way.

Fixed: 106086
Change-Id: I8f55276fdc030cf4e5d7b84903d7590acf3c9def
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717344
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>