[fdf] Create bringup buildbot test group for DFv2
Create group `//bundles/buildbot:bringup_dfv2` that includes only
bringup tests that are passing when DFv2 is enabled. The goal is to make
existing build bots that are running DFv2 include this package instead
of `//bundles/buildbot:bringup` in order to get the bots passing and
monitored by the Fuchsia build gardeners in order to prevent regressions
in DFV2.

Bug: 106650
Change-Id: I09b57ca6e80a97fbae64059c5f64e1ebe0c6ae14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717627
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>