[build][fidl] Allowlist HLCPP without _hlcpp suffix
This CL adds an allowlist of targets that can depend on HLCPP bindings
without using the explicit "_hlcpp" suffix. The allowlist is very broad
but will be narrowed in later CLs.

This also starts out by adding the suffix in a few places that are hard
to migrate mechanically, e.g. in templates or where the label is
constructed by string manipulation.

As a side-effect, this fixes some cases (measure-tape,
dangerous_identifiers) where we were incorrectly depending on HLCPP
bindings when the JSON IR was all that was needed.

Bug: 108354
Change-Id: Ie398444988abb2390d95c0c6f65aa324b7cdc403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718410
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>