[audio][aml-g12-tdm] Add extra metadata consistency checks
Make sure that the number of channels enabled in lanes
matches the number of channels in the ring buffer, most
likely this is a configuration error if not true.
Also only default bits per sample and slot for I2S and
StereoLeftJustified DAI types, making it mandatory for
TDM cases to explicitly specify these sizes.

Test: device-enumeration-test on Nelson

Change-Id: I7f15f46d4463d1627805e7f7a29d9ec9d8468480