[overnet] Fix router bug
The overnet router was keeping track of which nodes a route passed through,
but it was inserting the wrong node-id into the list.  As a result, if node
A had links to both B and C, it would think it could get to B both directly
_and_ via C, and similarly it would think it could get to C via B.  While
these loops should not cause actual problems routing (since the shortest path
is always used, and the indirect path is always slower than the direct path),
it is conceivable that in certain exception cases (as as links are shutting
down), some pathological behavior could be revealed. At the least, it's clearly
an error.

(Fixing this bug revealed an additional bug in one of the tests, one that had
been incorrectly passing.)

Change-Id: I87590ab9d9aa8c16f1750f727bdad5be0eb4b3d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717757
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>