[magma] Disable constrained mapping tests with HWASAN enabled
PlatformBuffer.MapConstrained and PlatformBuffer.Padding rely on mapping
a VMO into the first 4GB of address space, which may not be possible if
the HWASAN shadow stack is allocated there. So we need to disable the
tests when HWASAN is enabled.

Bug: 108382
Change-Id: I15e6bd0e705dcdb85341875f077704554357b403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719090
Reviewed-by: Leonard Chan <leonardchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>