[build] Add clang-doc templates and wrapper.
Creates some GN templates for invoking clang-doc.

Adds a script to synthesize a header file so the headers that should be
documented can be referenced by the build.

Adds a script to wrap a clang-doc invocation, redirecting the output to
a temporary directory that can be zipped (the build prohibits actions
writing output files that aren't known in advance).

I tested these manually but they are not currently hooked up. They will
be used in a followup by the tool that consumes the clang-doc output.

Change-Id: I563df5bacb6cb4a002a7bab23ce0d946f4513046
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711903
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>